SatplanExamples

Documentation for SatplanExamples.

SatplanExamples.get_fullsolution

SatplanExamples.get_moves

SatplanExamples.get_state

SatplanExamples.knights_problem

SatplanExamples.knights_problem_old

SatplanExamples.print_moves

SatplanExamples.print_state

SatplanExamples.sussman

SatplanExamples.to_DIMACS

SatplanExamples.get_fullsolution — Method

get_fullsolution(solution, codebook)

Extract the full solution from the SAT solver's output.

The solution is a plain text file with the SAT solver's output. The codebook is the one created when outputting the CNF to the DIMACS format.

SatplanExamples.get_moves — Method

get_moves(solution, codebook)

Extract moves from SAT solver's solution.

The solution is a plain text file with the SAT solver's output. The codebook is the one created when outputting the CNF to the DIMACS format.

SatplanExamples.get_state — Method

get_state(solution, codebook, turn)

Extract the SAT solver's solution's state at the indicated turn.

The solution is a plain text file with the SAT solver's output. The codebook is the one created when outputting the CNF to the DIMACS format.

SatplanExamples.knights_problem — Function

knights_problem(nr_ranks = 8, nr_files = 8, nr_moves = nr_files * nr_ranks, start_square = [1, 1]; open_tour = false)

Generates (more efficient) CNF representation of the knight's tour problem.

When generating the CNF representation for an open tour, set nr_moves to nr_files * nr_ranks - 1.

SatplanExamples.knights_problem_old — Function

knights_problem_old(nr_ranks = 8, nr_files = 8, nr_moves = nr_files * nr_ranks, start_square = [1, 1]; open_tour = false)

Generates (inefficient) CNF representation of the knight's tour problem.

When generating the CNF representation for an open tour, set nr_moves to nr_files * nr_ranks - 1.

SatplanExamples.print_moves — Method

print_moves(solution, codebook)

Print all moves from SAT solver's solution to the prompt.

The solution is a plain text file with the SAT solver's output. The codebook is the one created when outputting the CNF to the DIMACS format.

SatplanExamples.print_state — Method

print_state(solution, codebook, turn)

Print the SAT solver's solution's state at the indicated turn to the prompt.

The solution is a plain text file with the SAT solver's output. The codebook is the one created when outputting the CNF to the DIMACS format.

SatplanExamples.sussman — Function

sussman(nr_moves = 3)

Generate CNF representation of Sussman's anomaly, to be solved in the desired number of moves.

SatplanExamples.to_DIMACS — Method

to_DIMACS(cnf, file, codebook)

Exports set of clauses to its DIMACS representation.

The file will be fed to the SAT solver. The codebook is used during postprocessing.